Operator precedence
Since it is an == comparison it doesn't matter, but especially since this is educational, I suggest you add the parenthesis as the operator precedence isn't really obvious. Had it been an != comparison, the result would not have been the expected. Also, the code was spaced as if the & had higher precedence than ==.
http://en.cppreference.com/w/cpp/language/operator_precedence